refactor: share register name utilities
